refactor: 更改金额类型、POJO字段类型
将表示金额的类型改为 Long，
将 POJO 的基本类型改为包装类型

+++ boardgame-member/src/main/java/com/lxiaocode/boardgame/member/domain/Wallet.java
@@ -24,8 +24,8 @@ public class Wallet extends BaseEntity {
     private String memberId;
 
     @TableField("money")
-    private BigDecimal money;
+    private Long money;
 
     @TableField("froze")
-    private BigDecimal froze;
+    private Long froze;
 }
